+++ baby_respiration/app/__init__.py
@@ -1,4 +1,4 @@
 """Baby respiration detector package."""
 
-__version__ = "0.3.3"
+__version__ = "0.3.4"